+++ lib/view_component/base.rb
@@ -15,11 +15,21 @@ module ViewComponent
 
     ViewContextCalledBeforeRenderError = Class.new(StandardError)
 
+    RESERVED_PARAMETER = :content
+
     # For CSRF authenticity tokens in forms
     delegate :form_authenticity_token, :protect_against_forgery?, :config, to: :helpers
 
     class_attribute :content_areas
     self.content_areas = [] # class_attribute:default doesn't work until Rails 5.2
+
+    # EXPERIMENTAL: This API is experimental and may be removed at any time.
+    # Hook for allowing components to do work as part of the compilation process.
+    #
+    # For example, one might compile component-specific assets at this point.
+    def self._after_compile
+      # noop
+    end
 
     # Entrypoint for rendering components.
     #
@@ -68,8 +78,8 @@ module ViewComponent
       old_current_template = @current_template
       @current_template = self
 
-      # Assign captured content passed to component as a block to @content
-      @content = view_context.capture(self, &block) if block_given?
+      @_content_evaluated = false
+      @_render_in_block = block
 
       before_render
 
@@ -167,7 +177,20 @@ module ViewComponent
       @request ||= controller.request
     end
 
-    attr_reader :content, :view_context
+    attr_reader :view_context
+
+    def content
+      return @_content if defined?(@_content)
+      @_content_evaluated = true
+
+      @_content = if @view_context && @_render_in_block
+        view_context.capture(self, &@_render_in_block)
+      end
+    end
+
+    def content_evaluated?
+      @_content_evaluated
+    end
 
     # The controller used for testing components.
     # Defaults to ApplicationController. This should be set early
@@ -246,7 +269,14 @@ module ViewComponent
         if areas.include?(:content)
           raise ArgumentError.new ":content is a reserved content area name. Please use another name, such as ':body'"
         end
-        attr_reader(*areas)
+
+        areas.each do |area|
+          define_method area.to_sym do
+            content unless content_evaluated? # ensure content is loaded so content_areas will be defined
+            instance_variable_get(:"@#{area}") if instance_variable_defined?(:"@#{area}")
+          end
+        end
+
         self.content_areas = areas
       end
 
@@ -264,7 +294,7 @@ module ViewComponent
         parameter = validate_default ? collection_parameter : provided_collection_parameter
 
         return unless parameter
-        return if initialize_parameters.map(&:last).include?(parameter)
+        return if initialize_parameter_names.include?(parameter)
 
         # If Ruby cannot parse the component class, then the initalize
         # parameters will be empty and ViewComponent will not be able to render
@@ -281,7 +311,40 @@ module ViewComponent
         )
       end
 
+      # Ensure the component initializer does not define
+      # invalid parameters that could override the framework's
+      # methods.
+      def validate_initialization_parameters!
+        return unless initialize_parameter_names.include?(RESERVED_PARAMETER)
+
+        raise ArgumentError.new(
+          "#{self} initializer cannot contain " \
+          "`#{RESERVED_PARAMETER}` since it will override a " \
+          "public ViewComponent method."
+        )
+      end
+
+      def collection_parameter
+        if provided_collection_parameter
+          provided_collection_parameter
+        else
+          name && name.demodulize.underscore.chomp("_component").to_sym
+        end
+      end
+
+      def collection_counter_parameter
+        "#{collection_parameter}_counter".to_sym
+      end
+
+      def counter_argument_present?
+        instance_method(:initialize).parameters.map(&:second).include?(collection_counter_parameter)
+      end
+
       private
+
+      def initialize_parameter_names
+        initialize_parameters.map(&:last)
+      end
 
       def initialize_parameters
         instance_method(:initialize).parameters