+++ lib/view_component/collection.rb
@@ -8,7 +8,6 @@ module ViewComponent
     delegate :format, to: :component
 
     def render_in(view_context, &block)
-      component.compile(raise_errors: true)
       component.validate_collection_parameter!(validate_default: true)
 
       @collection.map do |item|