+++ test/app/components/slots_v2_before_render_component.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+class SlotsV2BeforeRenderComponent < ViewComponent::Base
+  include ViewComponent::SlotableV2
+
+  renders_one :title, "MyTitleComponent"
+  renders_many :greetings, "MyGreetingComponent"
+
+  class MyTitleComponent < ViewComponent::Base
+    attr_reader :prefix
+
+    def before_render
+      @prefix = "Testing"
+    end
+  end
+
+  class MyGreetingComponent < ViewComponent::Base
+    attr_reader :greeting
+
+    def before_render
+      @greeting = "Hello,"
+    end
+  end
+end